requirements.txt
arrow==1.2.3
bottle==0.12.23
plyer==2.1.0
psutil==5.9.4
PyQt5==5.15.7
pywin32==305
PyYAML==6.0
requests==2.28.1
